seeed/1-temperature-sensor/modules/temperature-sensor/requirements.txt
azure-iot-device
smbus2
RPi.bme280
requests
pymongo
azure-core==1.1.1
azure-cosmos==4.1.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
requests==2.22.0
six==1.13.0
urllib3==1.25.7
virtualenv==16.7.9
virtualenv-clone==0.5.3